Binary Saves
This replaces the Base64 encoded save format with a new binary one.
It's tiny. It's hard to modify. It shouldn't get in the way much -
you can break binary compatibility before release if you want, but
to preserve it, all you need to do is:

* Not remove properties from the save format
* Add new properties AFTER all existing ones
* Give every property the [Order] decorator/attribute

You don't need to touch the save/load code unless you're adding a
type it doesn't recognise.

Also, the SaveSystem provides the ReadSave() and WriteSave()
functions now, so other components of TimeHACK don't need to care
about the actual save format.